code clean up: LoginActviity.kt and activity_login.xml
