+++ app/src/main/AndroidManifest.xml
@@ -14,17 +14,21 @@
             android:exported="true">
         </activity>
         <activity
-            android:name=".SignUpActivity"
-            android:exported="true">
-        </activity>
-        <activity
-            android:name=".SignInActivity"
+            android:name=".LoginActivity"
             android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
 
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
+        </activity>
+        <activity
+            android:name=".SignUpActivity"
+            android:exported="true">
+        </activity>
+        <activity
+            android:name=".SignInActivity"
+            android:exported="true">
         </activity>
         <activity
             android:name=".MainActivity"
