+++ app/src/main/java/com/example/stow/MainActivity.kt
@@ -25,7 +25,7 @@ class MainActivity : AppCompatActivity() {
             // User is signed in
         } else {
             // No user is signed in
-            val intent = Intent(this, SignInActivity::class.java)
+            val intent = Intent(this, LoginActivity::class.java)
             startActivity(intent)
         }
         super.onCreate(savedInstanceState)
@@ -77,7 +77,7 @@ class MainActivity : AppCompatActivity() {
 
         binding.SignOutButton.setOnClickListener {
             FirebaseAuth.getInstance().signOut();
-            val intent = Intent(this, SignInActivity::class.java)
+            val intent = Intent(this, LoginActivity::class.java)
             startActivity(intent)
         }
