--- app/src/main/java/com/example/stow/SignInActivity.kt
@@ -1,54 +1,0 @@
-package com.example.stow
-
-import android.content.Intent
-import android.os.Bundle
-import android.widget.Toast
-import androidx.appcompat.app.AppCompatActivity
-import com.example.stow.databinding.ActivitySignInBinding
-import com.google.firebase.auth.FirebaseAuth
-
-class SignInActivity : AppCompatActivity() {
-
-    private lateinit var binding: ActivitySignInBinding
-    private lateinit var firebaseAuth: FirebaseAuth
-
-    override fun onCreate(savedInstanceState: Bundle?) {
-        val user = FirebaseAuth.getInstance().currentUser
-        if (user != null) {
-            // User is signed in
-            val intent = Intent(this, MainActivity::class.java)
-            startActivity(intent)
-        } else {
-            // No user is signed in
-        }
-        super.onCreate(savedInstanceState)
-
-        binding = ActivitySignInBinding.inflate(layoutInflater)
-        setContentView(binding.root)
-
-        firebaseAuth = FirebaseAuth.getInstance()
-
-        binding.login.setOnClickListener {
-            val email = binding.username.text.toString()
-            val pass = binding.password.text.toString()
-
-            if (email.isNotEmpty() && pass.isNotEmpty()) {
-                firebaseAuth.signInWithEmailAndPassword(email, pass).addOnCompleteListener {
-                    if (it.isSuccessful) {
-                        val intent = Intent(this, MainActivity::class.java)
-                        startActivity(intent)
-                    } else {
-                        Toast.makeText(this, it.exception.toString(), Toast.LENGTH_SHORT).show()
-                    }
-                }
-            } else {
-                Toast.makeText(this, "Empty Fields Are not Allowed !!", Toast.LENGTH_SHORT).show()
-            }
-        }
-
-        binding.SignUpButton.setOnClickListener {
-            val intent = Intent(this, SignUpActivity::class.java)
-            startActivity(intent)
-        }
-    }
-}
